pyproject.toml
[tool.poetry]
name = "watchsync"
version = "0.1.0"
description = "A Python-based tool for real-time file monitoring and synchronization, designed to replicate changes between local or remote locations."
authors = ["Roberto Lizana <rober.lizana@gmail.com>"]
readme = "README.md"
license = "AGPL-3.0"
homepage = "https://github.com/rlizana/watchsync"
repository = "https://github.com/rlizana/watchsync"
keywords = ["rsync", "watchdog", "real-time", "synchronization", "watchfiles", "sync", "inotify", "file monitoring", "file synchronization", "file replication", "file sync", "file watcher"]
packages = [
    { include = "watchsync", from = "." }
]


[tool.poetry.dependencies]
python = "^3.12"
cleo = "^2.1.0"
requests = "^2.32.3"
jinja2 = "^3.1.4"
watchfiles = "^1.0.3"
watchdog = "^6.0.0"
pyyaml = "^6.0.2"
coverage = "^7.6.10"

[tool.poetry.dev-dependencies]
pre-commit = "^3.0.0"
poetry-dynamic-versioning = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
watchsync = "watchsync.__main__:main"
watchsyncd = "watchsync.daemon.watchsyncd:main"

[tool.isort]
profile = "black"
known_first_party = ["watchsync"]
force_sort_within_sections = true
line_length = 79

[tool.flake8]
max-line-length = 79
import-order-style = "google"
application-import-names = ["watchsync"]
